+++ src/collective/contact/plonegroup/browser/tables.py
@@ -172,7 +172,7 @@ class DisplayGroupUsersView(BrowserView):
         if self.short:
             splitted = group_title.split('(')
             if len(splitted) > 1:
-                group_title = group_title.split('(')[-1][:-1]
+                group_title = group_title.split('(', 1)[-1][:-1]
         return html.escape(group_title)
 
     def _get_groups_and_members(self, group, index=0, keep_subgroups=False):
@@ -218,7 +218,9 @@ class DisplayGroupUsersView(BrowserView):
         for index, principal in self._get_groups_and_members(group, keep_subgroups=self.is_manager):
             # member may be a user or group
             isGroup = base_hasattr(principal, 'isGroup') and principal.isGroup() or 0
-            principal_title = html.escape(principal.getProperty('fullname') or principal.getProperty('title') or principal.getId())
+            principal_title = html.escape(principal.getProperty('fullname') or
+                                          principal.getProperty('title') or
+                                          principal.getId())
             if self.is_manager:
                 principal_title = principal_title + " ({0})".format(principal.id)
             principal_title = "<div class='user-or-group user-or-group-level-{0}'>{1}</div>".format(